requirements.txt
langchain
langchain-core
langchain-groq
python-dotenv
atlassian-python-api
slack_sdk
slack-bolt
websocket-client
notion-client
requests
streamlit
apscheduler
